+++ sample/miniinspector.html
@@ -2,7 +2,7 @@
 <html>
 	<head>
 			<meta charset="utf-8">
-			<title>Inspector sample</title>
+			<title>Mini inspector sample</title>
 			<script src="../node_modules/@ckeditor/ckeditor5-build-decoupled-document/build/ckeditor.js"></script>
 			<script src="../build/miniinspector.js"></script>
 	</head>
