+++ src/ckeditorinspector.js
@@ -20,7 +20,7 @@ import { updateViewState } from './view/data/actions';
 import { updateCommandsState } from './commands/data/actions';
 import EditorListener from './data/utils';
 
-import InspectorUI from './ui';
+import CKEditorInspectorUI from './ckeditorinspectorui';
 import Logger from './logger';
 import {
 	normalizeArguments,
@@ -280,7 +280,7 @@ export default class CKEditorInspector {
 
 		ReactDOM.render(
 			<Provider store={CKEditorInspector._store}>
-				<InspectorUI />
+				<CKEditorInspectorUI />
 			</Provider>,
 			container
 		);
