+++ src/ckeditorinspectorui.css
