+++ src/minickeditorinspector.js
@@ -8,7 +8,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import MiniInspectorUI from './miniinspectorui';
+import MiniInspectorUI from './minickeditorinspectorui';
 
 // From changelog -> webpack.
 window.CKEDITOR_MINI_INSPECTOR_VERSION = CKEDITOR_INSPECTOR_VERSION;
