+++ src/minickeditorinspectorui.css
